refactoring

+++ app/Http/Controllers/MatchingController.php
@@ -53,10 +53,8 @@ class MatchingController extends Controller
 
         $all_lehr = User::where('role', 'Lehr')->where('is_evaluable', true)->get();
         $all_stud = User::where('role', 'Stud')->where('is_evaluable', true)->get();
-        // $all_lehr = User::where('role', 'Lehr')->where('is_evaluable', true)->where('assigned', false)->get();
-        // $all_stud = User::where('role', 'Stud')->where('is_evaluable', true)->where('assigned', false)->get();
 
-        DB::table('lehr_stud_matchable')->truncate();
+        // DB::table('lehr_stud_matchable')->truncate();
 
         foreach ($all_lehr as $lehr) {
             $lehr->survey_data = json_decode($lehr->survey_data);
@@ -101,10 +99,11 @@ class MatchingController extends Controller
 
                             $mse = $this->mse($lehr, $stud);
                             $edge->setAttribute('graphviz.label', $mse . ', ');
-                            $lehr->matchable()->attach($stud, ['mse' => $mse]);
+                            // $lehr->matchable()->attach($stud, ['mse' => $mse]);
+                            $lehr->matchable()->syncWithoutDetaching([$stud->id => ['mse' => $mse]]);
 
-                            User::find($lehr->id)->update(['is_matchable' => true]);
-                            User::find($stud->id)->update(['is_matchable' => true]);
+                            // User::find($lehr->id)->update(['is_matchable' => true]);
+                            // User::find($stud->id)->update(['is_matchable' => true]);
                         } elseif (array_intersect($lehr->survey_data->faecher, $stud->survey_data->faecher)) {
                             $lehr_vertex = $graph->getVertex($lehr->id);
                             $stud_vertex = $graph->getVertex($stud->id);
@@ -113,10 +112,11 @@ class MatchingController extends Controller
 
                             $mse = $this->mse($lehr, $stud);
                             $edge->setAttribute('graphviz.label', $mse . ', ');
-                            $lehr->matchable()->attach($stud, ['mse' => $mse]);
+                            // $lehr->matchable()->attach($stud, ['mse' => $mse]);
+                            $lehr->matchable()->syncWithoutDetaching([$stud->id => ['mse' => $mse]]);
 
-                            User::find($lehr->id)->update(['is_matchable' => true]);
-                            User::find($stud->id)->update(['is_matchable' => true]);
+                            // User::find($lehr->id)->update(['is_matchable' => true]);
+                            // User::find($stud->id)->update(['is_matchable' => true]);
                         }
                     }
                 }
@@ -156,15 +156,17 @@ class MatchingController extends Controller
         $resultGraph->getVertex('s')->destroy();
         $resultGraph->getVertex('t')->destroy();
 
-        $unmatchable_users = User::where('is_evaluable', true)->where('is_matchable', false)->get();
-        foreach ($unmatchable_users as $user) {
-            if ($resultGraph->hasVertex($user->id)) {
-                $resultGraph->getVertex($user->id)->destroy();
+        $evaluable_users = User::where('is_evaluable', true)->get();
+        foreach ($evaluable_users as $user) {
+            if (!$user->is_matchable) {
+                if ($resultGraph->hasVertex($user->id)) {
+                    $resultGraph->getVertex($user->id)->destroy();
+                }
             }
         }
 
-        $matchable_lehr = User::where('role', 'Lehr')->where('is_matchable', true)->get();
-        $prematched_lehr = User::find(DB::table('lehr_stud')->pluck('lehr_id'));
+        $matchable_lehr = User::find(DB::table('lehr_stud')->pluck('lehr_id'));
+        $matched_lehr = User::find(DB::table('lehr_stud')->where('is_matched', true)->pluck('lehr_id'));
 
 
 
@@ -174,8 +176,8 @@ class MatchingController extends Controller
         // $graphviz->setFormat('svg');
         $graph_img = $graphviz->createImageHtml($resultGraph);
 
-        $prematched_graph = $resultGraph->createGraphClone();
-        foreach ($prematched_graph->getEdges() as $edge) {
+        $matched_graph = $resultGraph->createGraphClone();
+        foreach ($matched_graph->getEdges() as $edge) {
 
             $edge->setAttribute('graphviz.dir', 'none');
             $edge_lehr_id = $edge->getVertexStart()->getId();
@@ -183,15 +185,16 @@ class MatchingController extends Controller
 
             foreach ($matchable_lehr as $lehr) {
                 if ($edge_lehr_id == $lehr->id) {
-                    if ($lehr->assigned) {
-                        if ($lehr->prematched->first()->id == $edge_stud_id) {
+
+                    if ($lehr->matching_state == 'matched') {
+                        if ($lehr->matched_user->id == $edge_stud_id) {
                             $edge->setAttribute('graphviz.color', 'green');
                             $edge->setAttribute('graphviz.dir', 'both');
                         } else {
                             $edge->setAttribute('graphviz.color', 'red');
                         }
                     } else {
-                        if(User::find($edge_stud_id)->assigned) {
+                        if (User::find($edge_stud_id)->matching_state == 'matched') {
                             $edge->setAttribute('graphviz.color', 'red');
                         } else {
                             $edge->setAttribute('graphviz.color', 'black');
@@ -200,19 +203,146 @@ class MatchingController extends Controller
                 }
             }
         }
-        $prematched_graph->setAttribute('graphviz.graph.rankdir', 'LR');
-        $prematched_graph->setAttribute('graphviz.graph.bgcolor', 'transparent');
-        $prematched_graph_img = $graphviz->createImageHtml($prematched_graph);
+        $matched_graph->setAttribute('graphviz.graph.rankdir', 'LR');
+        $matched_graph->setAttribute('graphviz.graph.bgcolor', 'transparent');
+        $matched_graph_img = $graphviz->createImageHtml($matched_graph);
 
 
-        return view('matchable', compact('graph_img', 'prematched_graph_img', 'max_flow', 'prematched_lehr', 'matchable_lehr'));
+        return view('matchable', compact('graph_img', 'matched_graph_img', 'max_flow', 'matched_lehr', 'matchable_lehr'));
     }
+
+
+    public function setAssigned(Request $request, $lehrid, $studid)
+    {
+        $lehr = User::find($lehrid);
+        $stud = User::find($studid);
+
+        $lehr->matchable()->updateExistingPivot($stud, ['is_matched' => true]);
+
+        return back();
+    }
+
+
+    public function setUnassigned(Request $request, $lehrid, $studid)
+    {
+        $lehr = User::find($lehrid);
+        $stud = User::find($studid);
+
+        $lehr->matchable()->updateExistingPivot($stud, ['is_matched' => false]);
+
+        return back();
+    }
+
+
+    public function acceptMatching(Request $request)
+    {
+        $lehr = User::find($request->input('lehrid'));
+        $stud = User::find($request->input('studid'));
+        // $matching = DB::table('lehr_stud')->where('lehr_id', $request->input('lehrid'))->where('stud_id', $request->input('studid'))->first();
+        if ($request->input('role') == 'Lehr') {
+            // DB::table('lehr_stud')->where('lehr_id', $request->input('lehrid'))->where('stud_id', $request->input('studid'))->update([
+            //     'is_accepted_lehr' => true
+            // ]);
+            // $lehr->matchable()->updateExistingPivot($stud, ['is_accepted_lehr' => true]);
+            $lehr->matchable()->syncWithoutDetaching([$stud->id => ['is_accepted_lehr' => true]]);
+        }
+
+        if ($request->input('role') == 'Stud') {
+            // DB::table('lehr_stud')->where('lehr_id', $request->input('lehrid'))->where('stud_id', $request->input('studid'))->update([
+            //     'is_accepted_stud' => true
+            // ]);
+            // $stud->matchable()->updateExistingPivot($stud, ['is_accepted_stud' => true]);
+            $stud->matchable()->syncWithoutDetaching([$lehr->id => ['is_accepted_stud' => true]]);
+        }
+
+        return back();
+    }
+
+
+    public function declineMatching(Request $request)
+    {
+        $lehr = User::find($request->input('lehrid'));
+        $stud = User::find($request->input('studid'));
+        // $matching = DB::table('lehr_stud')->where('lehr_id', $request->input('lehrid'))->where('stud_id', $request->input('studid'))->first();
+        if ($request->input('role') == 'Lehr') {
+            // DB::table('lehr_stud')->where('lehr_id', $request->input('lehrid'))->where('stud_id', $request->input('studid'))->update([
+            //     'is_accepted_lehr' => false
+            // ]);
+            $lehr->matchable()->syncWithoutDetaching([$stud->id => ['is_accepted_lehr' => false]]);
+        }
+
+        if ($request->input('role') == 'Stud') {
+            // DB::table('lehr_stud')->where('lehr_id', $request->input('lehrid'))->where('stud_id', $request->input('studid'))->update([
+            //     'is_accepted_stud' => false
+            // ]);
+            $stud->matchable()->syncWithoutDetaching([$lehr->id => ['is_accepted_stud' => false]]);
+        }
+
+        return back();
+    }
+
+
+    public function notifyMatchings()
+    {
+
+        $unnotified_matchings = DB::table('lehr_stud')->where('is_notified', false)->get();
+
+        foreach ($unnotified_matchings as $unnotified_matching) {
+
+            try {
+                $lehr = User::find($unnotified_matching->lehr_id);
+                $lehr->notify(new MatchingProposal());
+            } catch (\Exception $e) {
+                $lehr->email_still_exists = false;
+                $lehr->save();
+            }
+
+            try {
+                $stud = User::find($unnotified_matching->stud_id);
+                $stud->notify(new MatchingProposal());
+            } catch (\Exception $e) {
+                $stud->email_still_exists = false;
+                $stud->save();
+            }
+
+
+            // DB::table('lehr_stud')->where('lehr_id', $lehr->id)->where('stud_id', $stud->id)->update(['created_at' => Carbon::now(), 'is_notified' => true]);
+            $lehr->matchable()->syncWithoutDetaching([$stud->id => ['is_notified' => true]]);
+        }
+
+        return back();
+    }
+
+
+    public function acceptedMatchings(Request $request)
+    {
+        $notified_matchings = DB::table('lehr_stud')->where('is_notified', true)->get();
+        foreach ($notified_matchings as $am) {
+            $am->lehr = User::find($am->lehr_id);
+            $am->stud = User::find($am->stud_id);
+            $am->elapsed_time = Carbon::parse($am->created_at)->diffForHumans(Carbon::now());
+        }
+
+        $accepted_matchings = DB::table('lehr_stud')->where('is_accepted_lehr', true)->where('is_accepted_stud', true)->get();
+        foreach ($accepted_matchings as $am) {
+            $am->lehr = User::find($am->lehr_id);
+            $am->stud = User::find($am->stud_id);
+        }
+        return view('accepted_matchings', ['notified_matchings' => $notified_matchings, 'accepted_matchings' => $accepted_matchings]);
+    }
+
+
+
+
+
+
+
 
 
     public function matchings(Request $request)
     {
-        $lehr = User::where('role', 'lehr')->where('is_evaluable', true)->where('assigned', false)->get();
-        $stud = User::where('role', 'stud')->where('is_evaluable', true)->where('assigned', false)->get();
+        $lehr = User::where('role', 'lehr')->where('is_evaluable', true)->where('matching_state', 'unmatched')->get();
+        $stud = User::where('role', 'stud')->where('is_evaluable', true)->where('matching_state', 'unmatched')->get();
 
         foreach ($stud as $current_stud) {
             $current_stud->survey_data = json_decode($current_stud->survey_data);
@@ -290,141 +420,12 @@ class MatchingController extends Controller
         $assigned_matchings = DB::table('lehr_stud')->where('is_notified', false)->get();
         $assigned = [];
         foreach ($assigned_matchings as $am) {
-            $assigned_lehr = User::where('role', 'lehr')->where('is_evaluable', true)->where('assigned', true)->where('id', $am->lehr_id)->get();
-            $assigned_stud = User::where('role', 'stud')->where('is_evaluable', true)->where('assigned', true)->where('id', $am->stud_id)->get();
+            $assigned_lehr = User::where('role', 'lehr')->where('is_evaluable', true)->where('matching_state', 'prematched')->where('id', $am->lehr_id)->get();
+            $assigned_stud = User::where('role', 'stud')->where('is_evaluable', true)->where('matching_state', 'prematched')->where('id', $am->stud_id)->get();
             $assigned[] = ['lehr' => $assigned_lehr[0], 'stud' => $assigned_stud[0], 'is_accepted_lehr' => $am->is_accepted_lehr, 'is_accepted_stud' => $am->is_accepted_stud, 'mse' => $am->mse, 'elapsed_time' => Carbon::parse($am->created_at)->diffForHumans(Carbon::now())];
         }
         // dd($assigned_matchings);
 
         return view('matchings', ['users' => $lehr, 'matchings' => $matchings, 'assigned_matchings' => $assigned, 'unmatchable_lehr' => $unmatchable_lehr]);
     }
-
-    public function setAssigned(Request $request, $lehrid, $studid, $mse)
-    {
-        $lehr = User::where('role', 'lehr')->where('is_evaluable', true)->where('assigned', false)->where('id', $lehrid)->first();
-        $stud = User::where('role', 'stud')->where('is_evaluable', true)->where('assigned', false)->where('id', $studid)->first();
-
-        $lehr->matchings()->attach($stud, ['mse' => $mse]);
-        // DB::insert('insert into lehr_stud (lehr_id, stud_id, mse) values (?, ?, ?)', [$lehrid, $studid, $mse]);
-
-        $lehr->assigned = true;
-        $lehr->save();
-
-        $stud->assigned = true;
-        $stud->save();
-
-        return back();
-    }
-
-    public function setUnassigned(Request $request, $lehrid, $studid)
-    {
-        $lehr = User::where('role', 'lehr')->where('is_evaluable', true)->where('assigned', true)->where('id', $lehrid)->first();
-        $stud = User::where('role', 'stud')->where('is_evaluable', true)->where('assigned', true)->where('id', $studid)->first();
-
-        $lehr->assigned = false;
-        $lehr->save();
-
-        $stud->assigned = false;
-        $stud->save();
-
-        DB::table('lehr_stud')->where('lehr_id', $lehrid)->where('stud_id', $studid)->delete();
-
-        return back();
-    }
-
-
-    public function acceptMatching(Request $request)
-    {
-        // $matching = DB::table('lehr_stud')->where('lehr_id', $request->input('lehrid'))->where('stud_id', $request->input('studid'))->first();
-        if ($request->input('role') == 'Lehr') {
-            DB::table('lehr_stud')->where('lehr_id', $request->input('lehrid'))->where('stud_id', $request->input('studid'))->update([
-                'is_accepted_lehr' => true
-            ]);
-        }
-
-        if ($request->input('role') == 'Stud') {
-            DB::table('lehr_stud')->where('lehr_id', $request->input('lehrid'))->where('stud_id', $request->input('studid'))->update([
-                'is_accepted_stud' => true
-            ]);
-        }
-
-        $matching = DB::table('lehr_stud')->where('lehr_id', $request->input('lehrid'))->where('stud_id', $request->input('studid'))->first();
-        if ($matching->is_accepted_lehr && $matching->is_accepted_stud) {
-        }
-
-        return back();
-    }
-
-
-    public function declineMatching(Request $request)
-    {
-        // $matching = DB::table('lehr_stud')->where('lehr_id', $request->input('lehrid'))->where('stud_id', $request->input('studid'))->first();
-        if ($request->input('role') == 'Lehr') {
-            DB::table('lehr_stud')->where('lehr_id', $request->input('lehrid'))->where('stud_id', $request->input('studid'))->update([
-                'is_accepted_lehr' => false
-            ]);
-        }
-
-        if ($request->input('role') == 'Stud') {
-            DB::table('lehr_stud')->where('lehr_id', $request->input('lehrid'))->where('stud_id', $request->input('studid'))->update([
-                'is_accepted_stud' => false
-            ]);
-        }
-
-        $matching = DB::table('lehr_stud')->where('lehr_id', $request->input('lehrid'))->where('stud_id', $request->input('studid'))->first();
-        if ($matching->is_accepted_lehr == 0 ||  $matching->is_accepted_stud == 0) {
-        }
-
-        return back();
-    }
-
-
-    public function notifyMatchings()
-    {
-
-        $unnotified_matchings = DB::table('lehr_stud')->where('is_notified', false)->get();
-
-        foreach ($unnotified_matchings as $unnotified_matching) {
-
-            try {
-                $lehr = User::find($unnotified_matching->lehr_id);
-                $lehr->notify(new MatchingProposal());
-            } catch(\Exception $e) {
-                $lehr->email_still_exists = false;
-                $lehr->save();
-            }
-
-            try {
-                $stud = User::find($unnotified_matching->stud_id);
-                $stud->notify(new MatchingProposal());
-            } catch(\Exception $e) {
-                $stud->email_still_exists = false;
-                $stud->save();
-            }
-
-
-            DB::table('lehr_stud')->where('lehr_id', $lehr->id)->where('stud_id', $stud->id)->update(['created_at' => Carbon::now(), 'is_notified' => true]);
-        }
-
-        return back();
-    }
-
-
-    public function acceptedMatchings(Request $request)
-    {
-        $notified_matchings = DB::table('lehr_stud')->where('is_notified', true)->get();
-        foreach ($notified_matchings as $am) {
-            $am->lehr = User::find($am->lehr_id);
-            $am->stud = User::find($am->stud_id);
-            $am->elapsed_time = Carbon::parse($am->created_at)->diffForHumans(Carbon::now());
-        }
-
-        $accepted_matchings = DB::table('lehr_stud')->where('is_accepted_lehr', true)->where('is_accepted_stud', true)->get();
-        foreach ($accepted_matchings as $am) {
-            $am->lehr = User::find($am->lehr_id);
-            $am->stud = User::find($am->stud_id);
-        }
-        return view('accepted_matchings', ['notified_matchings' => $notified_matchings, 'accepted_matchings' => $accepted_matchings]);
-    }
-
 }
